+++ silo-core/test/invariants/CryticToFoundry.t.sol
@@ -477,6 +477,7 @@ contract CryticToFoundry is Invariants, Setup {
         echidna_BASE_INVARIANT();
     }
 
+    // FOUNDRY_PROFILE=core_test forge test -vv --ffi --mt test_replayechidna_LENDING_INVARIANT
     function test_replayechidna_LENDING_INVARIANT() public {
         Tester.deposit(1, 0, 0, 1);
         echidna_LENDING_INVARIANT();